fix: authentication flow error when it's referenced as an IdP first broker login

+++ src/main/java/de/adorsys/keycloak/config/service/AuthenticationFlowsImportService.java
@@ -26,11 +26,13 @@ import de.adorsys.keycloak.config.model.RealmImport;
 import de.adorsys.keycloak.config.properties.ImportConfigProperties;
 import de.adorsys.keycloak.config.properties.ImportConfigProperties.ImportManagedProperties.ImportManagedPropertiesValues;
 import de.adorsys.keycloak.config.repository.AuthenticationFlowRepository;
+import de.adorsys.keycloak.config.repository.IdentityProviderRepository;
 import de.adorsys.keycloak.config.repository.RealmRepository;
 import de.adorsys.keycloak.config.util.AuthenticationFlowUtil;
 import de.adorsys.keycloak.config.util.CloneUtil;
 import org.keycloak.representations.idm.AuthenticationExecutionInfoRepresentation;
 import org.keycloak.representations.idm.AuthenticationFlowRepresentation;
+import org.keycloak.representations.idm.IdentityProviderRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -61,6 +63,7 @@ public class AuthenticationFlowsImportService {
     private final ExecutionFlowsImportService executionFlowsImportService;
     private final AuthenticatorConfigImportService authenticatorConfigImportService;
     private final UsedAuthenticationFlowWorkaroundFactory workaroundFactory;
+    private final IdentityProviderRepository identityProviderRepository;
 
     private final ImportConfigProperties importConfigProperties;
 
@@ -70,7 +73,8 @@ public class AuthenticationFlowsImportService {
             AuthenticationFlowRepository authenticationFlowRepository,
             ExecutionFlowsImportService executionFlowsImportService,
             AuthenticatorConfigImportService authenticatorConfigImportService, UsedAuthenticationFlowWorkaroundFactory workaroundFactory,
-            ImportConfigProperties importConfigProperties
+            ImportConfigProperties importConfigProperties,
+            IdentityProviderRepository identityProviderRepository
     ) {
         this.realmRepository = realmRepository;
         this.authenticationFlowRepository = authenticationFlowRepository;
@@ -78,6 +82,7 @@ public class AuthenticationFlowsImportService {
         this.authenticatorConfigImportService = authenticatorConfigImportService;
         this.workaroundFactory = workaroundFactory;
         this.importConfigProperties = importConfigProperties;
+        this.identityProviderRepository = identityProviderRepository;
     }
 
     /**
@@ -330,6 +335,24 @@ public class AuthenticationFlowsImportService {
         workaround.resetFlowIfNeeded();
     }
 
+
+    /**
+     * Returns true if the flow is referenced by an identity provider in the given realm,
+     * either as first broker login flow or as post broker login flow.
+     *
+     * @param realmName the keycloak realm name
+     * @param flowAlias the alias of the flow to check
+     * @return true if the flow is referenced, false if not
+     */
+
+    private boolean isFlowReferencedByIdP(String realmName, String flowAlias) {
+        List<IdentityProviderRepresentation> idps = identityProviderRepository.getAll(realmName);
+        return idps.stream().anyMatch(idp ->
+                flowAlias.equals(idp.getFirstBrokerLoginFlowAlias())
+                        || flowAlias.equals(idp.getPostBrokerLoginFlowAlias())
+        );
+    }
+
     private void deleteTopLevelFlowsMissingInImport(
             RealmImport realmImport,
             List<AuthenticationFlowRepresentation> importedTopLevelFlows
@@ -345,8 +368,12 @@ public class AuthenticationFlowsImportService {
         for (AuthenticationFlowRepresentation existingTopLevelFlow : existingTopLevelFlows) {
             if (topLevelFlowsToImportAliases.contains(existingTopLevelFlow.getAlias())) continue;
 
-            logger.debug("Delete authentication flow: {}", existingTopLevelFlow.getAlias());
-            authenticationFlowRepository.delete(realmName, existingTopLevelFlow.getId());
+            if (!isFlowReferencedByIdP(realmName, existingTopLevelFlow.getAlias())) {
+                logger.debug("Delete authentication flow: {}", existingTopLevelFlow.getAlias());
+                authenticationFlowRepository.delete(realmName, existingTopLevelFlow.getId());
+            } else {
+                logger.warn("Cannot delete authentication flow '{}' as it is referenced by an identity provider", existingTopLevelFlow.getAlias());
+            }
         }
     }
 }
